alot

+++ DalApi/Order.cs
@@ -12,7 +12,7 @@ namespace DO
         public uint HostingUnitKey { get; set; }
         public uint GuestRequestKey { get; set; }
         public string HostId { get; set; }
-        public float Fee  { get; set; }
+        public double Fee  { get; set; }
         public OrderStatus Status { get; set; }
         public DateTime OrderDate { get; set; }
         public DateTime SentDate { get; set; }
